refactor(client): changed structure of utils
