+++ client/src/lib/staticData.ts
