+++ client/src/pages/[set]/write.tsx
@@ -7,9 +7,9 @@ import { setApi } from 'apis/setApi';
 import Custom404 from 'pages/404';
 import { Button } from 'ui/Button';
 import { CardInterface, SetInterface } from 'interfaces';
-import { isAnswerCorrect } from 'utils/isAnswerCorrect';
+import { isAnswerCorrect } from 'lib/utils';
 import { useLocalStore } from 'storage/useLocalStore';
-import { isBackendLess } from 'utils/staticData';
+import { isBackendLess } from 'lib/staticData';
 import Header from 'ui/Header';
 import style from 'styles/pages/write.module.scss';
