+++ client/src/ui/Layout.tsx
@@ -5,8 +5,8 @@ import { useQuery } from 'react-query';
 import { AnimatePresence, motion } from 'framer-motion';
 import { authApi } from 'apis/authApi';
 import { useUserStore } from 'storage/useUserStore';
-import { localMemory, sessionMemory } from 'utils/browserMemory';
-import { growUpMotions, isBackendLess, pathsForHidingLayout } from 'utils/staticData';
+import { localMemory, sessionMemory } from 'lib/browserMemory';
+import { growUpMotions, isBackendLess, pathsForHidingLayout } from 'lib/staticData';
 import { FolderEditing } from 'modules/FolderEditing';
 import { Button } from 'ui/Button';
 import Header from 'ui/Header';
