--- client/src/utils/fontSizeBasedOnLength.ts
@@ -1,4 +1,0 @@
-export const fontSizeBasedOnLength = (length: number) => {
-  if (length > 100) return '1.25rem';
-  else return '1.5rem';
-};
